Extracted ViewState to the class
