+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/payments/cardreader/onboarding/CardReaderOnboardingViewModelTest.kt
@@ -30,16 +30,15 @@ import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboa
 import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingState.WcpayNotActivated
 import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingState.WcpayNotInstalled
 import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingEvent
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.CashOnDeliveryDisabledState
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.GenericErrorState
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.LoadingState
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.NoConnectionErrorState
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.StripeAcountError
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.StripeExtensionError
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.UnsupportedErrorState
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.WCPayError
-import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewModel.OnboardingViewState.WCPayError.WCPayNotInstalledState
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.CashOnDeliveryDisabledState
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.GenericErrorState
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.LoadingState
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.NoConnectionErrorState
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.StripeAcountError
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.StripeExtensionError
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.UnsupportedErrorState
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.WCPayError
+import com.woocommerce.android.ui.payments.cardreader.onboarding.CardReaderOnboardingViewState.WCPayError.WCPayNotInstalledState
 import com.woocommerce.android.ui.payments.cardreader.onboarding.PluginType.STRIPE_EXTENSION_GATEWAY
 import com.woocommerce.android.ui.payments.cardreader.onboarding.PluginType.WOOCOMMERCE_PAYMENTS
 import com.woocommerce.android.viewmodel.BaseUnitTest
@@ -1597,7 +1596,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
             val viewModel = createVM()
 
             assertThat(viewModel.viewStateData.value).isInstanceOf(
-                OnboardingViewState.SelectPaymentPluginState::class.java
+                CardReaderOnboardingViewState.SelectPaymentPluginState::class.java
             )
         }
 
@@ -1610,7 +1609,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             assertThat(viewStateData.cardIllustration).isNotNull()
         }
 
@@ -1623,7 +1622,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             assertThat(viewStateData.headerLabel).isNotNull()
         }
 
@@ -1636,7 +1635,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             assertThat(viewStateData.choosePluginHintLabel).isNotNull()
         }
 
@@ -1649,7 +1648,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             assertThat(viewStateData.selectWcPayButtonLabel).isNotNull
         }
 
@@ -1662,7 +1661,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             assertThat(viewStateData.selectStripeButtonLabel).isNotNull
         }
 
@@ -1675,7 +1674,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             assertThat(viewStateData.confirmPaymentMethodButtonLabel).isNotNull()
         }
 
@@ -1689,7 +1688,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
             val viewModel = createVM()
 
             val state = (
-                viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+                viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
                 )
             assertThat(state.cardIllustration)
                 .describedAs("Check illustration")
@@ -1752,7 +1751,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             viewStateData.onConfirmPaymentMethodClicked.invoke(WOOCOMMERCE_PAYMENTS)
 
             assertThat(viewModel.viewStateData.value).isEqualTo(LoadingState)
@@ -1768,7 +1767,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             viewStateData.onConfirmPaymentMethodClicked.invoke(WOOCOMMERCE_PAYMENTS)
 
             verify(appPrefsWrapper).removeLastConnectedCardReaderId()
@@ -1784,7 +1783,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             viewStateData.onConfirmPaymentMethodClicked.invoke(WOOCOMMERCE_PAYMENTS)
 
             verify(cardReaderManager).disconnectReader()
@@ -1824,7 +1823,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
 
             val viewModel = createVM()
 
-            val viewStateData = viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState
+            val viewStateData = viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState
             viewStateData.onConfirmPaymentMethodClicked.invoke(STRIPE_EXTENSION_GATEWAY)
 
             assertThat(viewModel.viewStateData.value).isEqualTo(LoadingState)
@@ -1837,7 +1836,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
                 WcpayNotActivated
             )
             val viewModel = createVM()
-            val receivedViewStates = mutableListOf<OnboardingViewState>()
+            val receivedViewStates = mutableListOf<CardReaderOnboardingViewState>()
             viewModel.viewStateData.observeForever {
                 receivedViewStates.add(it)
             }
@@ -2295,7 +2294,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
                 .thenReturn(ChoosePaymentGatewayProvider)
 
             val viewModel = createVM()
-            (viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState)
+            (viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState)
                 .onConfirmPaymentMethodClicked.invoke(WOOCOMMERCE_PAYMENTS)
 
             verify(tracker).trackPaymentGatewaySelected(WOOCOMMERCE_PAYMENTS)
@@ -2308,7 +2307,7 @@ class CardReaderOnboardingViewModelTest : BaseUnitTest() {
                 .thenReturn(ChoosePaymentGatewayProvider)
 
             val viewModel = createVM()
-            (viewModel.viewStateData.value as OnboardingViewState.SelectPaymentPluginState)
+            (viewModel.viewStateData.value as CardReaderOnboardingViewState.SelectPaymentPluginState)
                 .onConfirmPaymentMethodClicked.invoke(STRIPE_EXTENSION_GATEWAY)
 
             verify(tracker).trackPaymentGatewaySelected(STRIPE_EXTENSION_GATEWAY)
